Product images fixes for cart and product details page. #35

+++ database/migrations/1644541931786_api_tokens.ts
@@ -5,9 +5,9 @@ export default class ApiTokens extends BaseSchema {
 
   public async up () {
     this.schema.createTable(this.tableName, (table) => {
-      table.increments('id').primary()
+      table.bigIncrements('id').primary()
 
-      table.integer('user_id').unsigned().references('id').inTable('users').onDelete('CASCADE')
+      table.bigInteger('user_id').unsigned().references('id').inTable('users').onDelete('CASCADE')
 
       table.string('name').notNullable()
